+++ ChangeLog.markdown
@@ -8,9 +8,9 @@ PHPUnit_Selenium 1.2.4
 
 * Supporting form submit (also via children elements).
 * Supporting radio boxes.
-* Supporting implicit waits.
-* Supporting back button.
-* Supporting refresh of pages.
+* Supporting implicit waits on $this->by*().
+* Supporting back and forward buttons via $this->back() and $this->forward().
+* Supporting refresh of pages via $this->refresh().
 * Correctly marking Selenium 1 tests as skipped when server is not running.
 
 PHPUnit_Selenium 1.2.3