+++ ChangeLog.markdown
@@ -3,6 +3,11 @@ PHPUnit_Selenium 1.2
 
 This is the list of changes for the PHPUnit_Selenium 1.2 release series.
 
+PHPUnit_Selenium 1.2.5
+----------------------
+
+* Implemented $this->timeouts()->asyncScript().
+
 PHPUnit_Selenium 1.2.4
 ----------------------